p2p:refactor p2pnext as system p2p plugin dht
