+++ system/init.go
@@ -6,11 +6,10 @@
 package system
 
 import (
-	_ "github.com/33cn/chain33/p2p/gossip" // init p2p
-	_ "github.com/33cn/chain33/p2pnext"
 	_ "github.com/33cn/chain33/system/consensus/init" //register consensus init package
 	_ "github.com/33cn/chain33/system/crypto/init"
 	_ "github.com/33cn/chain33/system/dapp/init"
 	_ "github.com/33cn/chain33/system/mempool/init"
+	_ "github.com/33cn/chain33/system/p2p/init" // init p2p plugin
 	_ "github.com/33cn/chain33/system/store/init"
 )
