+++ system/p2p/dht/net/mdns.go
@@ -1,4 +1,4 @@
-package dht
+package net
 
 import (
 	"context"
