+++ system/p2p/dht/protocol/broadcast/broadcast.go
@@ -13,12 +13,12 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/peer"
 
-	prototypes "github.com/33cn/chain33/p2pnext/protocol/types"
+	prototypes "github.com/33cn/chain33/system/p2p/dht/protocol/types"
 	core "github.com/libp2p/go-libp2p-core"
 
 	"github.com/33cn/chain33/common/log/log15"
-	p2pty "github.com/33cn/chain33/p2pnext/types"
 	"github.com/33cn/chain33/queue"
+	p2pty "github.com/33cn/chain33/system/p2p/dht/types"
 	"github.com/33cn/chain33/types"
 )
