+++ system/p2p/dht/protocol/broadcast/broadcast_test.go
@@ -6,13 +6,14 @@ package broadcast
 
 import (
 	"encoding/hex"
-	"github.com/33cn/chain33/p2p"
 	"testing"
 
+	"github.com/33cn/chain33/p2p"
+
 	"github.com/33cn/chain33/client"
-	prototypes "github.com/33cn/chain33/p2pnext/protocol/types"
-	p2pty "github.com/33cn/chain33/p2pnext/types"
 	"github.com/33cn/chain33/queue"
+	prototypes "github.com/33cn/chain33/system/p2p/dht/protocol/types"
+	p2pty "github.com/33cn/chain33/system/p2p/dht/types"
 	"github.com/33cn/chain33/types"
 	"github.com/stretchr/testify/assert"
 )
@@ -39,7 +40,7 @@ var (
 
 func newTestEnv(q queue.Queue) *prototypes.P2PEnv {
 
-	cfg := types.NewChain33Config(types.ReadFile("../../../cmd/chain33/chain33.test.toml"))
+	cfg := types.NewChain33Config(types.ReadFile("../../../../../cmd/chain33/chain33.test.toml"))
 	q.SetConfig(cfg)
 	go q.Start()
