+++ system/p2p/dht/types/const.go
@@ -1,0 +1,13 @@
+// Copyright Fuzamei Corp. 2018 All Rights Reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+package types
+
+const (
+
+	// DHTTypeName p2p插件名称，底层基于libp2p框架, dht结构化网络
+	DHTTypeName = "dht"
+
+	// DefaultP2PPort 默认端口
+	DefaultP2PPort = 13803
+)
