+++ util/cli/chain33.go
@@ -15,12 +15,13 @@ package cli
 import (
 	"flag"
 	"fmt"
-	"github.com/33cn/chain33/p2p"
 	"net/http"
 	_ "net/http/pprof" //
 	"os"
 	"path/filepath"
 	"runtime"
+
+	"github.com/33cn/chain33/p2p"
 
 	"github.com/33cn/chain33/metrics"
